refactor(server-core): share the JWKS certificate field builder
buildCertificateJwkFields was duplicated verbatim in the JwkController and
the realm jwks handlers. Hoist it into core/key/certificate.ts alongside
buildX5c/buildX5tS256 so the fail-open x5c/x5t#S256 derivation has a single
source of truth.
