+++ apps/server-core/src/adapters/http/controllers/workflows/jwks/module.ts
@@ -19,32 +19,12 @@ import type { IAppEvent } from 'routup';
 import type { Repository } from 'typeorm';
 import { In } from 'typeorm';
 import type { KeyEntity } from '../../../../database/domains/index.ts';
-import { buildX5c, buildX5tS256, parseCertificateChain } from '../../../../../core/index.ts';
+import { buildCertificateJwkFields } from '../../../../../core/index.ts';
 import { getRequestStringParam } from '../../../request/index.ts';
 
 export type JwkControllerContext = {
     repository: Repository<KeyEntity>
 };
-
-async function buildCertificateJwkFields(
-    certificate: string | null,
-): Promise<Partial<Pick<OAuth2JsonWebKey, 'x5c' | 'x5t#S256'>>> {
-    if (!certificate) {
-        return {};
-    }
-
-    try {
-        const chain = parseCertificateChain(certificate);
-        return {
-            x5c: buildX5c(chain),
-            'x5t#S256': await buildX5tS256(chain),
-        };
-    } catch {
-        // One malformed legacy/database row must not take down the realm's
-        // whole JWKS. Publish the usable public key without certificate data.
-        return {};
-    }
-}
 
 @DController('')
 export class JwkController {
